hooked up agent to db and also set up inputting the api key
