+++ web-app/mlclient.py
@@ -11,11 +11,16 @@ import json
 import traceback
 import requests
 
-# pylint ignore
+# pylint: disable=E0401
 
 from llama_index.core.tools import FunctionTool
 from llama_index.core.agent import ReActAgent
 from llama_index.llms.openai import OpenAI
+
+def set_api_key(key):
+    """This sets the environment API key"""
+    os.environ["OPENAI_API_KEY"] = str(key)
+
 
 BASE_MEALDB_API_URL = "https://www.themealdb.com/api/json/v1/1/"
 
@@ -60,45 +65,41 @@ def lookup_recipe_by_id(meal_id: str) -> dict:
         return {
             "error": f"Debug lookup_recipie_by_id tool Error: Invalid meal_id provided: {meal_id}"
         }
-    return call_mealdb_api("lookup.php", {"i", meal_id})
+    return call_mealdb_api("lookup.php", {"i": meal_id})
 
 
 # LlamaIndex part
+def initialize_agent():
+    """ initialize the LlamaIndex ReAct agent """
 
-# Tools that are being provided to the agent
+    ingredient_tool = FunctionTool.from_defaults(
+        fn=filter_by_ingredient,
+        name="filter_by_ingredient",
+        description="""Use this tool FIRST to get a list of meal names and IDs based on a single primary ingredient
+    This provides options but not full recipes""",
+    )
 
-ingredient_tool = FunctionTool.from_defaults(
-    fn=filter_by_ingredient,
-    name="filter_by_ingredient",
-    description="""Use this tool FIRST to get a list of meal names and IDs based on a single primary ingredient
-This provides options but not full recipes""",
-)
-
-recipe_lookup_tool = FunctionTool.from_defaults(
-    fn=lookup_recipe_by_id,
-    name="lookup_recipe_by_id",
-    description="""Use this tool SECOND, provide a specific numeric meal ID (obtained from 'filter_by_ingredient' tool)
-to get the full recipe details, including ingredients and instructions""",
-)
+    recipe_lookup_tool = FunctionTool.from_defaults(
+        fn=lookup_recipe_by_id,
+        name="lookup_recipe_by_id",
+        description="""Use this tool SECOND, provide a specific numeric meal ID (obtained from 'filter_by_ingredient' tool)
+    to get the full recipe details, including ingredients and instructions""",
+    )
 
 
-category_tool = FunctionTool.from_defaults(
-    fn=lambda category: call_mealdb_api("filter.php", {"c": category}),
-    name="filter_by_category",
-    description="""Finds a LIST of meals by category (such as 'Seafood', 'Vegetarian') proides options but not full recipes""",
-)
+    # LLM Setup
+    try:
+        llm = OpenAI(model="gpt-4.1-mini")  # This model seems to work
+    except Exception as e:
+        print(f"CRITICAL: Error initializing LLM. check OPENAI_API_KEY set? {e}")
+        sys.exit(1)
 
-# LLM Setup
-try:
-    llm = OpenAI(model="gpt-4o-mini")  # This model seems to work
-except Exception as e:
-    print(f"CRITICAL: Error initializing LLM. check OPENAI_API_KEY set? {e}")
-    sys.exit(1)
+    # ReAct Agent setup
+    agent = ReActAgent.from_tools(
+        tools=[ingredient_tool, recipe_lookup_tool], llm=llm, verbose=True
+    )
 
-# ReAct Agent setup
-agent = ReActAgent.from_tools(
-    tools=[ingredient_tool, recipe_lookup_tool, category_tool], llm=llm, verbose=True
-)
+    return agent
 
 # getting the agent output
 
@@ -116,12 +117,75 @@ def print_agent_output(output_data: Dict[str, Any]):
     print(final_response if final_response else "No final response captured.")
 
 
+def get_recipe_reccomendation(user_input) -> Dict[str, Any]:
+    """Process user input and return recipe recommendations using the LlamaIndex agent, runs the thing"""
+    agent = initialize_agent()
+    stdout_capture_buffer = io.StringIO()
+    captured_stdout_log = None
+    agent_response = None
+
+    with redirect_stdout(stdout_capture_buffer):
+        try:
+            agent_query = (
+                f"User query: '{user_input}'."
+                """#YOUR GOAL: Based on the user's ingredients and preferences,
+                FIRST find relevant meal IDs using 'filter_by_ingredient'.
+                THEN select the best matching meal ID based on
+                user preferences (if any). NEXT, use 'lookup_recipe_by-id' with the selected ID to get the full recipe.
+                FINALLY extract and return the cooking instructions (from 'strInstructions'), and some comments wishing the user well
+                for their goals. If no specific preference beyond ingredient is given, you can choose one, make a comment about how/why you think the
+                user might like it.
+                """
+            )
+            # Execute the agent
+            agent_response = agent.chat(agent_query)
+
+        except Exception as e:
+            print("\n Error during agent execution")
+            print(f"The error is : {e}")
+            agent_response = f"Agent execution failed: {str(e)}"
+
+    captured_stdout_log = stdout_capture_buffer.getvalue()
+
+    agent_output_dict = {
+        "recipe": str(agent_response),
+        "reasons": captured_stdout_log,
+        "inputs": user_input
+    }
+
+    agent_output_string = " "
+
+    agent_output_string += "----------Recipe----------\n"
+    agent_output_string += agent_output_dict["recipe"]
+    agent_output_string += "\n" + "\n"
+    agent_output_string += "-----------Why We Chose This Recipe For You----------\n"
+    agent_output_string += "-----------(our AI's agent's thoughts and internet accesses)----------\n"
+    agent_output_string += agent_output_dict["reasons"]
+    agent_output_string += "\n" + "\n"
+    agent_output_string += "-----------Given these inputs----------\n"
+    agent_output_string += agent_output_dict["inputs"]
+
+
+    return agent_output_string
+
+
+
 # test, will need to find out how to integrate this into the app
 
-user_input = input(
-    "Enter some ingredients, a main ingredient, and a style of dish such as 'continental'"
-)
+if __name__ == "__main__":
+    user_input = input(
+        "Enter some ingredients, a main ingredient, and a style of dish such as 'continental'"
+    )
 
+    result = get_recipe_reccomendation(user_input)
+    print(result.get("verbose_log_and_tool_prints", "No stdout"))
+    print(result.get("final_agent_response", "No final response captured."))
+
+
+# pylint: disable=W0105
+
+'''
+reference from old main:
 stdout_capture_buffer = io.StringIO()
 captured_stdout_log = None
 agent_response = None
@@ -144,16 +208,18 @@ with redirect_stdout(stdout_capture_buffer):
         agent_response = agent.chat(agent_query)
 
     except Exception as e:
-        print(f"\n Error during agent execution")
+        print("\n Error during agent execution")
         print(f"The error is : {e}")
         agent_response = "Agent execution failed see cap. logs."
 
 captured_stdout_log = stdout_capture_buffer.getvalue()
 
-agent_output_dict = {
+agent_output_dict_test = {
     "verbose_log_and_tool_prints": captured_stdout_log,
     "final_agent_response": agent_response,
 }
 
 # This part is just debug
-print_agent_output(agent_output_dict)
+print_agent_output(agent_output_dict_test)
+
+'''
