+++ web-app/templates/index.html
@@ -18,6 +18,19 @@
             <h1 class="text-3xl font-semibold text-stone-800 mb-6 text-center">Personal Recipe Generator</h1>
             <form id="recipeForm" class="space-y-6">
                 <div>
+                    <label for="apiKey" class="block text-sm font-medium text-stone-700 mb-2">
+                        OpenAI API Key
+                    </label>
+                    <input
+                        type="password"
+                        id="apiKey"
+                        name="apiKey"
+                        class="w-full px-4 py-2 border border-stone-300 rounded-md focus:ring-2 focus:ring-stone-500 focus:border-stone-500"
+                        placeholder="Enter your OpenAI API key (sk-...)"
+                        required
+                    />
+                </div>
+                <div>
                     <label for="ingredients" class="block text-sm font-medium text-stone-700 mb-2">
                         What ingredients do you have?
                     </label>
@@ -96,7 +109,7 @@
 
         document.getElementById('recipeForm').addEventListener('submit', async (e) => {
             e.preventDefault();
-            
+
             const ingredients = document.getElementById('ingredients').value;
             const goals = document.getElementById('goals').value;
             const additionalNotes = document.getElementById('additional_notes').value;
@@ -104,7 +117,8 @@
             const data = {
                 ingredients: ingredients,
                 goals: goals,
-                additionalNotes: additionalNotes
+                additionalNotes: additionalNotes,
+                apiKey: document.getElementById('apiKey').value
             };
 
             try {
@@ -115,9 +129,9 @@
                     },
                     body: JSON.stringify(data)
                 });
-                
+
                 const result = await response.json();
-                
+
                 // Create recipe object
                 const recipe = {
                     title: `Recipe for ${ingredients.split('\n')[0]}...`,
@@ -143,4 +157,4 @@
         });
     </script>
 </body>
-</html> 
+</html>
